provide Win32 native threading
Summary:
Add an implementation for the Win32 threading model as a backing API for
the internal c++ threading interfaces.  This uses the Fls* family for
the TLS (which has the support for adding termination callbacks),
CRITICAL_SECTIONs for the recursive mutex, and Slim Reader/Writer locks
(SRW locks) for non-recursive mutexes.  These APIs should all be
available on Vista or newer.

Reviewers: mclow.lists, EricWF

Subscribers: halyavin, awson, kimgr, cfe-commits, kastiglione, smeenai, majnemer, rnk

Differential Revision: https://reviews.llvm.org/D28220