Explicitly specify MSVC mangling of iostream globals
Summary:
With MS ABI, the `char []` definitions of the iostream globals (`iostream.cpp`)
result in mangled symbol names that don't match the mangled symbol names of the
declarations (`<iostream>`).

For example, `istream cin` mangles to:

    ?cin@__1@std@@3V?$basic_istream@DU?$char_traits@D@__1@std@@@12@A

but its definition in iostream.cpp is `char cin[sizeof(istream)]`, which
incorrectly mangles to:

    ?cin@__1@std@@3PADA

To work aound this issue, this change manually sets the symbol name using an
`__asm__` label, for win32 only.

Reviewers: mclow.lists, compnerd, EricWF

Subscribers: smeenai, majnemer, llvm-commits

Differential Revision: https://reviews.llvm.org/D23378